+++ DataFirst/CarPool.Data.Models/VehicleDBO.cs
@@ -16,7 +16,7 @@ namespace CarPool.Data.Models
         public int UserID { get; set; }
         public string Maker { get; set; }
         public string Number { get; set; }
-        public byte Seats { get; set; }
+        public int Seats { get; set; }
         public bool IsActive { get; set; }
         public VehicleType Type { get; set; }      
         public virtual UserDBO User { get; set; }
