+++ DataFirst/CarPool.Models/Offer.cs
@@ -12,7 +12,7 @@ namespace CarPool.Application.Models
         public int Destination { get; set; }
         public string StartDate { get; set; }
         public string EndDate { get; set; }
-        public byte SeatsAvailable { get; set; }
+        public int SeatsAvailable { get; set; }
         public List<ViaPoints> ViaPoints { get; set; }      
     }
 }
