+++ DataFirst/CarPool.Models/Vehicle.cs
@@ -6,7 +6,7 @@
         public int UserID { get; set; }
         public string Maker { get; set; }
         public string Number { get; set; }
-        public byte Seats { get; set; }
+        public int Seats { get; set; }
         public bool IsActive { get; set; }
         public int Type { get; set; }      
     }
